refactor: `TextIoConsoleTest`를 입출력에 따른 `TextIoInputTest`와 `TextIoOutputTest`로 분리
