+++ src/test/java/com/programmers/voucher/global/io/textio/TextIoInputTest.java
@@ -1,14 +1,7 @@
-package com.programmers.voucher.global.io;
+package com.programmers.voucher.global.io.textio;
 
-import com.programmers.voucher.domain.customer.domain.Customer;
-import com.programmers.voucher.domain.voucher.domain.FixedAmountVoucher;
-import com.programmers.voucher.domain.voucher.domain.PercentDiscountVoucher;
-import com.programmers.voucher.domain.voucher.domain.Voucher;
 import com.programmers.voucher.domain.voucher.domain.VoucherType;
 import com.programmers.voucher.domain.voucher.dto.request.VoucherCreateRequest;
-import com.programmers.voucher.global.io.textio.TextIoConsole;
-import com.programmers.voucher.global.io.textio.TextIoInput;
-import com.programmers.voucher.global.io.textio.TextIoOutput;
 import org.beryx.textio.TextIO;
 import org.beryx.textio.mock.MockTextTerminal;
 import org.junit.jupiter.api.BeforeEach;
@@ -18,18 +11,15 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import java.util.List;
-import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
 
-class TextIoConsoleTest {
-
+class TextIoInputTest {
     private MockTextTerminal mockTextTerminal;
     private TextIO textIO;
 
-    private TextIoConsole textIoConsole;
     private TextIoInput textIoInput;
-    private TextIoOutput textIoOutput;
 
     @BeforeEach
     void init() {
@@ -37,11 +27,7 @@ class TextIoConsoleTest {
         textIO = new TextIO(mockTextTerminal);
 
         textIoInput = new TextIoInput(textIO);
-        textIoOutput = new TextIoOutput(textIO);
-
-        textIoConsole = new TextIoConsole(textIoInput, textIoOutput);
     }
-
 
     @Test
     @DisplayName("성공: 바우처 생성 정보 입력 실행 - 바우처 타입(fixed)")
@@ -52,7 +38,7 @@ class TextIoConsoleTest {
         inputs.add("10");
 
         //when
-        VoucherCreateRequest result = textIoConsole.inputVoucherCreateInfo();
+        VoucherCreateRequest result = textIoInput.inputVoucherCreateInfo();
 
         //then
         assertThat(result.getVoucherType()).isEqualTo(VoucherType.FIXED_AMOUNT);
@@ -68,7 +54,7 @@ class TextIoConsoleTest {
         inputs.add("10");
 
         //when
-        VoucherCreateRequest result = textIoConsole.inputVoucherCreateInfo();
+        VoucherCreateRequest result = textIoInput.inputVoucherCreateInfo();
 
         //then
         assertThat(result.getVoucherType()).isEqualTo(VoucherType.PERCENT);
@@ -85,7 +71,7 @@ class TextIoConsoleTest {
         inputs.add("10");
 
         //when
-        VoucherCreateRequest result = textIoConsole.inputVoucherCreateInfo();
+        VoucherCreateRequest result = textIoInput.inputVoucherCreateInfo();
 
         //then
         assertThat(result.getVoucherType()).isEqualTo(VoucherType.FIXED_AMOUNT);
@@ -105,7 +91,7 @@ class TextIoConsoleTest {
         inputs.add("10");
 
         //when
-        VoucherCreateRequest result = textIoConsole.inputVoucherCreateInfo();
+        VoucherCreateRequest result = textIoInput.inputVoucherCreateInfo();
 
         //then
         assertThat(result.getVoucherType()).isEqualTo(VoucherType.FIXED_AMOUNT);
@@ -125,45 +111,10 @@ class TextIoConsoleTest {
         inputs.add("10");
 
         //when
-        VoucherCreateRequest result = textIoConsole.inputVoucherCreateInfo();
+        VoucherCreateRequest result = textIoInput.inputVoucherCreateInfo();
 
         //then
         assertThat(result.getVoucherType()).isEqualTo(VoucherType.PERCENT);
         assertThat(result.getAmount()).isEqualTo(10);
-    }
-
-    @Test
-    @DisplayName("성공 - 회원 목록 출력")
-    void printCustomers() {
-        //given
-        Customer customerA = new Customer(UUID.randomUUID(), "customerA");
-        Customer customerB = new Customer(UUID.randomUUID(), "customerB");
-        List<Customer> givenCustomers = List.of(customerA, customerB);
-
-        //when
-        textIoConsole.printCustomers(givenCustomers);
-
-        //then
-        String expectedOutput = customerA.fullInfoString() + "\n" + customerB.fullInfoString();
-        String output = mockTextTerminal.getOutput();
-
-        assertThat(output).isEqualTo(expectedOutput);
-    }
-
-    @Test
-    void printVouchers() {
-        //given
-        FixedAmountVoucher fixedVoucher = new FixedAmountVoucher(UUID.randomUUID(), 10);
-        PercentDiscountVoucher percentVoucher = new PercentDiscountVoucher(UUID.randomUUID(), 10);
-        List<Voucher> givenVouchers = List.of(fixedVoucher, percentVoucher);
-
-        //when
-        textIoConsole.printVouchers(givenVouchers);
-
-        //then
-        String expectedOutput = fixedVoucher.fullInfoString() + "\n" + percentVoucher.fullInfoString();
-        String output = mockTextTerminal.getOutput();
-
-        assertThat(output).isEqualTo(expectedOutput);
     }
 }
